refactor: sortMissionEntities関数を分離してコード整理
sortMissionEntitiesをindexから独立したファイルに切り出し
テストファイルも対応して移動・修正
