+++ src/app/(private)/_components/mission/entitity/index.tsx
@@ -1,15 +1,6 @@
-import { ENTITY_ORDER } from '@/app/_utils/constants';
 import { css } from '@/styled-system/css';
 import { EntityIcon, type MissionEntity } from '../../entity-icon';
-
-export const sortMissionEntities: (
-  a: MissionEntity,
-  b: MissionEntity,
-) => number = (a, b) => {
-  const typeA = ENTITY_ORDER.indexOf(a.itemType);
-  const typeB = ENTITY_ORDER.indexOf(b.itemType);
-  return typeA - typeB;
-};
+import { sortMissionEntities } from './sort-mission-entities';
 
 export const MissionEntities = (props: { items: MissionEntity[] }) => {
   return (
